refactor(scheduler): 重构任务优先级类型和调度器实现
将TaskPriority从u8类型重构为具有命名常量的新类型，提高可读性和类型安全性
优化调度器实现中的统计计算和资源管理，减少可变借用冲突
移除未使用的演示二进制文件配置，简化Cargo.toml

+++ 03-embedded-concepts/projects/scheduler/src/analysis/mod.rs
@@ -2,7 +2,9 @@
 
 use crate::{Scheduler, SchedulerResult, SchedulerStatistics};
 use core::sync::atomic::{AtomicU32, Ordering};
+use core::cell::RefCell;
 use alloc::vec::Vec;
+use alloc::format;
 
 /// 实时性能分析器
 pub struct PerformanceAnalyzer {
@@ -88,7 +90,7 @@ impl PerformanceAnalyzer {
             return 0.0;
         }
         
-        let sum: f64 = history.iter().map(|s| s.cpu_utilization).sum();
+        let sum: f64 = history.iter().map(|s| s.cpu_utilization as f64).sum();
         sum / history.len() as f64
     }
